Make all analysis fields optional
It seems the spotify API has changed their responses from the audio_analysis endpoint.

A number of fields seem to be missing and/or have different types now
- some numbers are now contained in strings e.g. `timestamp`.

I've gone the heavy handed route and made all the fields optional. I
started trying to only make certain fields optional but this was hit and
miss as I tried other tracks.

Perhaps there is a better solution?