Fix cached token scopes
## Fix

Serde rename "scopes" to "scope", but leave struct type as "scopes".

## Info

The cached auth token was not being restored from the cache, prompting the login flow every time.

On inspecting the cache, the "scopes" property was empty.

Turns out this was because the form response from spotify does not contain a "scopes" field. But instead contains a "scope".

This PR simply renames the property to "scopes" as it's a small change. Let me know if there's a more appropriate fix.